+++ src/core_modules/capture-core/components/WidgetProfile/WidgetProfile.component.tsx
@@ -15,6 +15,7 @@ import {
     useProgram,
     useTrackedEntityInstances,
     useClientAttributesWithSubvalues,
+    useUserRoles,
     useTeiDisplayName,
     useUserRoles,
 } from './hooks';
@@ -98,12 +99,12 @@ const WidgetProfilePlain = ({
         programWriteAccess,
         trackedEntityTypeWriteAccess,
     } = useEnrollmentAccessContext();
-    const trackedEntityTypeName = program?.trackedEntityType?.displayName;
     const {
         loading: userRolesLoading,
         error: userRolesError,
         userRoles,
     } = useUserRoles();
+    const trackedEntityTypeName = program?.trackedEntityType?.displayName;
 
     const hasNoAttributes = !program?.programTrackedEntityAttributes?.length;
 
@@ -114,7 +115,7 @@ const WidgetProfilePlain = ({
 
     const profileButtonLabel = useMemo(() => {
         if (readOnlyMode || hasNoAttributes) return null;
-        if (!isEditable) return i18n.t('Show profile');
+        if (!isEditable) return i18n.t('View profile');
         if (isEditable) return i18n.t('Edit');
         return null;
     }, [isEditable, readOnlyMode, hasNoAttributes]);