+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [106.3.0](https://github.com/dhis2/capture-app/compare/v106.2.8...v106.3.0) (2026-06-02)
+
+
+### Features
+
+* [DHIS2-21371] Read-only enrollment dashboard ([#4529](https://github.com/dhis2/capture-app/issues/4529)) ([55e0def](https://github.com/dhis2/capture-app/commit/55e0def7f2ae4ffd0dd2d3602afbba0f7e8c03da))
+
+## [106.2.8](https://github.com/dhis2/capture-app/compare/v106.2.7...v106.2.8) (2026-06-01)
+
+
+### Bug Fixes
+
+* [DHIS2-21567] String without spaces overflows form ([#4584](https://github.com/dhis2/capture-app/issues/4584)) ([9dc1fc8](https://github.com/dhis2/capture-app/commit/9dc1fc8698fdbe7f330ccc271c95ffc2a3c6b9f5))
+
 ## [106.2.7](https://github.com/dhis2/capture-app/compare/v106.2.6...v106.2.7) (2026-05-26)